+++ src/main/java/duke/EventTask.java
@@ -1,27 +1,29 @@
+package duke;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
 /**
- * The  DeadlineTask class is a Task that takes a deadline.
+ * The EventTask is a Task that has at event time.
  */
-public class DeadlineTask extends Task {
+public class EventTask extends Task {
 
     protected static final String COLON = ":";
 
     protected LocalDate date;
     protected LocalTime time;
+
     /**
-     * Constructor for a DeadlineTask object.
-     * @param description    Name of the task.
-     * @param by             Date of the deadline.
+     * Constructor of a EventTask object
+     * @param description      Name of the task
+     * @param at               Date of the event.
      */
-    public DeadlineTask(String description, String by) throws DukeException {
+    public EventTask(String description, String at) throws DukeException {
         super(description);
-        String[] dateParts = by.split(" ");
+        String[] dateParts = at.split(" ");
         if (dateParts.length != 2) {
-            throw new DukeException("Deadline Date and Time is specified wrongly");
+            throw new DukeException("Event Date and Time is specified wrongly");
         }
         this.date = LocalDate.parse(dateParts[0]);
         String temp = "";
@@ -37,12 +39,11 @@ public class DeadlineTask extends Task {
     public String dateTimeString() {
         return this.date.toString() + " " + this.time.toString().replace(":", "");
     }
-
     @Override
     public String toString() {
         DateTimeFormatter formatterDate = DateTimeFormatter.ofPattern("MMM dd yyyy");
         DateTimeFormatter formatterTime =DateTimeFormatter.ofPattern("h:mm a");
-        return "[D]" + super.toString() + " (by: "
+        return "[E]" + super.toString() + " (at: "
                 + this.date.format(formatterDate) + " "
                 + this.time.format(formatterTime) + ")";
     }
