+++ src/main/java/duke/Ui.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.io.IOException;
 import java.time.format.DateTimeParseException;
 import java.util.Scanner;
